refactor: 파일명 변경 및 어댑터 코드 개선
