--- app/src/main/java/com/eighteen/eighteenandroid/presentation/ViewPagerAdapter.kt
@@ -1,31 +1,0 @@
-package com.eighteen.eighteenandroid.presentation
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.eighteen.eighteenandroid.R
-
-class ViewPagerAdapter (
-    private var list: List<String>,
-) : RecyclerView.Adapter<ViewPagerAdapter.AdapterViewHolder>() {
-
-    override fun getItemCount(): Int {
-        return list.size
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AdapterViewHolder {
-        val v = LayoutInflater.from(parent.context).inflate(R.layout.profile_item, parent, false)
-        return AdapterViewHolder(v)
-    }
-
-    override fun onBindViewHolder(holder: AdapterViewHolder, position: Int) {
-        val item = list[position]
-        holder.title.text = item
-    }
-
-    inner class AdapterViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        val title: TextView = view.findViewById(R.id.tx_profile_name)
-    }
-}
