+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/backgroundColor"
-    tools:context=".presentation.MainFragment">
+    tools:context=".presentation.home.MainFragment">
 
     <TextView
         android:id="@+id/tx_today_teen"
@@ -23,7 +23,7 @@
         app:layout_constraintStart_toStartOf="parent"/>
 
     <com.google.android.material.chip.ChipGroup
-        android:id="@+id/tag_group"
+        android:id="@+id/chip_group"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
@@ -38,13 +38,13 @@
         app:chipCornerRadius="16dp" />
 
     <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/profile_view_pager"
+        android:id="@+id/vp_today_teen"
         android:layout_width="269dp"
         android:layout_height="346dp"
         android:layout_margin="20dp"
         android:orientation="horizontal"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tag_group"/>
+        app:layout_constraintTop_toBottomOf="@id/chip_group"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
