+++ app/src/main/res/layout/today_teen_item.xml
@@ -13,7 +13,7 @@
         android:layout_margin="8dp">
 
         <TextView
-            android:id="@+id/tx_profile_name"
+            android:id="@+id/tx_name"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:text="Profile"
